fixbug3

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Services\UserAdminService;
 use App\Models\User;
+use App\Models\Auction;
+use App\Models\Bid;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
@@ -126,7 +128,19 @@ class UserController extends Controller
      */
     public function destroy($userId)
     {
-        $user = User::where('user_id', $userId)->delete();
-        return redirect()->route('listUser')->with('message','削除しました！');
+        $userSelling = Auction::where('selling_user_id',  $userId)
+            ->get()
+            ->toArray();
+
+        $userBid = Bid::where('user_id', $userId)
+            ->get()
+            ->toArray();
+
+        if ($userSelling || $userBid) {
+            return redirect()->back()->with('warning', 'アカウントがアクテイブしています');
+        } else {
+            $user = User::where('user_id', $userId)->delete();
+            return redirect()->route('listUser')->with('message','削除しました！');
+        }
     }
 }
